move enum + license
